ScalaPsiUtil cleanup: move some functions to InferUtil Remove accidentally committed code Apply all inspections for modified files
