+++ src/org/jetbrains/plugins/scala/lang/psi/ScalaPsiUtil.scala
@@ -2,6 +2,8 @@ package org.jetbrains.plugins.scala
 package lang
 package psi
 
+import java.util.concurrent.ConcurrentMap
+
 import com.intellij.codeInsight.PsiEquivalenceUtil
 import com.intellij.lang.java.JavaLanguage
 import com.intellij.openapi.diagnostic.Logger
@@ -19,7 +21,7 @@ import com.intellij.psi.search.{GlobalSearchScope, LocalSearchScope, SearchScope
 import com.intellij.psi.stubs.StubElement
 import com.intellij.psi.tree.TokenSet
 import com.intellij.psi.util._
-import com.intellij.util.containers.ConcurrentWeakHashMap
+import com.intellij.util.containers.ContainerUtil
 import org.jetbrains.plugins.scala.caches.CachesUtil
 import org.jetbrains.plugins.scala.editor.typedHandler.ScalaTypedHandler
 import org.jetbrains.plugins.scala.extensions._
@@ -64,7 +66,6 @@ import scala.annotation.tailrec
 import scala.collection.mutable.ArrayBuffer
 import scala.collection.{Seq, Set, mutable}
 import scala.reflect.NameTransformer
-import scala.util.control.ControlThrowable
 
 /**
  * User: Alexander Podkhalyuzin
@@ -132,8 +133,8 @@ object ScalaPsiUtil {
                 contains(annot.typeElement.getText.replace(" ", ""))
       }
     } else {
-      s.hasAnnotation("scala.reflect.BooleanBeanProperty") != None ||
-        s.hasAnnotation("scala.beans.BooleanBeanProperty") != None
+      s.hasAnnotation("scala.reflect.BooleanBeanProperty").isDefined ||
+        s.hasAnnotation("scala.beans.BooleanBeanProperty").isDefined
     }
   }
 
@@ -145,8 +146,8 @@ object ScalaPsiUtil {
                 contains(annot.typeElement.getText.replace(" ", ""))
       }
     } else {
-      s.hasAnnotation("scala.reflect.BeanProperty") != None ||
-        s.hasAnnotation("scala.beans.BeanProperty") != None
+      s.hasAnnotation("scala.reflect.BeanProperty").isDefined ||
+        s.hasAnnotation("scala.beans.BeanProperty").isDefined
     }
   }
 
@@ -207,7 +208,6 @@ object ScalaPsiUtil {
     s match {
       case Seq() =>
         // object A { def foo(a: Any) = ()}; A foo () ==>> A.foo(()), or A.foo() ==>> A.foo( () )
-        val unitExpr = ScalaPsiElementFactory.createExpressionFromText("()", manager)
         Some(Seq(new Expression(types.Unit, place)))
       case _ =>
         val exprTypes: Seq[ScType] =
@@ -252,7 +252,7 @@ object ScalaPsiUtil {
   }
 
   /**
-  Pick all type parameters by method maps them to the appropriate type arguments, if they are
+  *Pick all type parameters by method maps them to the appropriate type arguments, if they are
    */
   def inferMethodTypesArgs(fun: PsiMethod, classSubst: ScSubstitutor): ScSubstitutor = {
 
@@ -350,7 +350,7 @@ object ScalaPsiUtil {
       } else if (implicitMap.isEmpty) checkImplicits(secondPart = false, noApplicability = true)
     }
     if (implicitMap.isEmpty) None
-    else if (implicitMap.length == 1) Some(implicitMap.apply(0))
+    else if (implicitMap.length == 1) Some(implicitMap.head)
     else MostSpecificUtil(ref, 1).mostSpecificForImplicit(implicitMap.toSet)
   }
 
@@ -359,6 +359,7 @@ object ScalaPsiUtil {
     Option[ImplicitResolveResult] = {
     val oldAlg = ScalaProjectSettings.getInstance(e.getProject).isUseOldImplicitConversionAlg
     if (oldAlg) {
+      //noinspection ScalaDeprecation
       return oldFindImplicitConversion(e, refName, ref, processor, noImplicitsForArgs)
     }
     lazy val funType = Option(
@@ -467,10 +468,10 @@ object ScalaPsiUtil {
       } else if (implicitMap.isEmpty) checkImplicits(secondPart = false, noApplicability = true)
     }
     if (implicitMap.length == 1) {
-      val rr = implicitMap.apply(0)
+      val rr = implicitMap.head
       specialExtractParameterType(rr) match {
         case Some(ScFunctionType(tp, _)) =>
-          Some(ImplicitResolveResult(tp, rr.getElement, rr.importsUsed, rr.substitutor, ScSubstitutor.empty, false)) //todo: from companion parameter
+          Some(ImplicitResolveResult(tp, rr.getElement, rr.importsUsed, rr.substitutor, ScSubstitutor.empty, isFromCompanion = false)) //todo: from companion parameter
         case _ => None
       }
     } else None
@@ -644,11 +645,11 @@ object ScalaPsiUtil {
   @tailrec
   def isAnonymousExpression(expr: ScExpression): (Int, ScExpression) = {
     val seq = ScUnderScoreSectionUtil.underscores(expr)
-    if (seq.length > 0) return (seq.length, expr)
+    if (seq.nonEmpty) return (seq.length, expr)
     expr match {
       case b: ScBlockExpr =>
         if (b.statements.length != 1) (-1, expr)
-        else if (b.lastExpr == None) (-1, expr)
+        else if (b.lastExpr.isEmpty) (-1, expr)
         else isAnonymousExpression(b.lastExpr.get)
       case p: ScParenthesisedExpr => p.expr match {case Some(x) => isAnonymousExpression(x) case _ => (-1, expr)}
       case f: ScFunctionExpr =>  (f.parameters.length, expr)
@@ -663,8 +664,8 @@ object ScalaPsiUtil {
     index.getModuleForFile(element.getContainingFile.getVirtualFile)
   }
 
-  val collectImplicitObjectsCache: ConcurrentWeakHashMap[(ScType, Project, GlobalSearchScope), Seq[ScType]] =
-    new ConcurrentWeakHashMap()
+  val collectImplicitObjectsCache: ConcurrentMap[(ScType, Project, GlobalSearchScope), Seq[ScType]] =
+    ContainerUtil.newConcurrentMap[(ScType, Project, GlobalSearchScope), Seq[ScType]]()
 
   def collectImplicitObjects(_tp: ScType, project: Project, scope: GlobalSearchScope): Seq[ScType] = {
     val tp = ScType.removeAliasDefinitions(_tp)
@@ -756,20 +757,6 @@ object ScalaPsiUtil {
       }
     }
 
-    case class MyType()
-    object MyImplicitTypeHelper {
-      class MyImplicitType{}
-
-      implicit def typeToImplicitType(in: MyType): MyImplicitType = ???
-    }
-    import MyImplicitTypeHelper._
-
-    def myFunc: MyImplicitType = {
-      val myType: MyType = MyType()
-      myType
-    }
-
-
     collectParts(tp)
     val res: mutable.HashMap[String, Seq[ScType]] = new mutable.HashMap
     def addResult(fqn: String, tp: ScType): Unit = {
@@ -786,6 +773,7 @@ object ScalaPsiUtil {
       val part = parts.dequeue()
       //here we want to convert projection types to right projections
       val visited = new mutable.HashSet[PsiClass]()
+      @tailrec
       def collectObjects(tp: ScType) {
         tp match {
           case types.Any =>
@@ -920,207 +908,6 @@ object ScalaPsiUtil {
 
   def getSettings(project: Project): ScalaCodeStyleSettings = {
     CodeStyleSettingsManager.getSettings(project).getCustomSettings(classOf[ScalaCodeStyleSettings])
-  }
-
-  def undefineSubstitutor(typeParams: Seq[TypeParameter]): ScSubstitutor = {
-    typeParams.foldLeft(ScSubstitutor.empty) {
-      (subst: ScSubstitutor, tp: TypeParameter) =>
-        subst.bindT((tp.name, ScalaPsiUtil.getPsiElementId(tp.ptp)),
-          new ScUndefinedType(new ScTypeParameterType(tp.ptp, ScSubstitutor.empty)))
-    }
-  }
-
-  def localTypeInference(retType: ScType, params: Seq[Parameter], exprs: Seq[Expression],
-                         typeParams: Seq[TypeParameter],
-                         shouldUndefineParameters: Boolean = true,
-                         safeCheck: Boolean = false,
-                         filterTypeParams: Boolean = true): ScTypePolymorphicType = {
-    localTypeInferenceWithApplicability(retType, params, exprs, typeParams, shouldUndefineParameters, safeCheck,
-      filterTypeParams)._1
-  }
-
-
-  class SafeCheckException extends ControlThrowable
-
-  //todo: move to InferUtil
-  def localTypeInferenceWithApplicability(retType: ScType, params: Seq[Parameter], exprs: Seq[Expression],
-                                          typeParams: Seq[TypeParameter],
-                                          shouldUndefineParameters: Boolean = true,
-                                          safeCheck: Boolean = false,
-                                          filterTypeParams: Boolean = true): (ScTypePolymorphicType, Seq[ApplicabilityProblem]) = {
-    val (tp, problems, _, _) = localTypeInferenceWithApplicabilityExt(retType, params, exprs, typeParams,
-      shouldUndefineParameters, safeCheck, filterTypeParams)
-    (tp, problems)
-  }
-
-  def localTypeInferenceWithApplicabilityExt(retType: ScType, params: Seq[Parameter], exprs: Seq[Expression],
-                                             typeParams: Seq[TypeParameter],
-                                             shouldUndefineParameters: Boolean = true,
-                                             safeCheck: Boolean = false,
-                                             filterTypeParams: Boolean = true
-    ): (ScTypePolymorphicType, Seq[ApplicabilityProblem], Seq[(Parameter, ScExpression)], Seq[(Parameter, ScType)]) = {
-    // See SCL-3052, SCL-3058
-    // This corresponds to use of `isCompatible` in `Infer#methTypeArgs` in scalac, where `isCompatible` uses `weak_<:<`
-    val s: ScSubstitutor = if (shouldUndefineParameters) undefineSubstitutor(typeParams) else ScSubstitutor.empty
-    val abstractSubst = ScTypePolymorphicType(retType, typeParams).abstractTypeSubstitutor
-    val paramsWithUndefTypes = params.map(p => p.copy(paramType = s.subst(p.paramType),
-      expectedType = abstractSubst.subst(p.paramType)))
-    val c = Compatibility.checkConformanceExt(checkNames = true, paramsWithUndefTypes, exprs, checkWithImplicits = true,
-      isShapesResolve = false)
-    val tpe = if (c.problems.isEmpty) {
-      var un: ScUndefinedSubstitutor = c.undefSubst
-      val subst = c.undefSubst
-      subst.getSubstitutor(!safeCheck) match {
-        case Some(unSubst) =>
-          if (!filterTypeParams) {
-            val undefiningSubstitutor = new ScSubstitutor(typeParams.map(typeParam => {
-              ((typeParam.name, ScalaPsiUtil.getPsiElementId(typeParam.ptp)), new ScUndefinedType(new ScTypeParameterType(typeParam.ptp, ScSubstitutor.empty)))
-            }).toMap, Map.empty, None)
-            ScTypePolymorphicType(retType, typeParams.map(tp => {
-              var lower = tp.lowerType()
-              var upper = tp.upperType()
-              def hasRecursiveTypeParameters(typez: ScType): Boolean = {
-                var hasRecursiveTypeParameters = false
-                typez.recursiveUpdate {
-                  case tpt: ScTypeParameterType =>
-                    typeParams.find(tp => (tp.name, ScalaPsiUtil.getPsiElementId(tp.ptp)) ==(tpt.name, tpt.getId)) match {
-                      case None => (true, tpt)
-                      case _ =>
-                        hasRecursiveTypeParameters = true
-                        (true, tpt)
-                    }
-                  case tp: ScType => (hasRecursiveTypeParameters, tp)
-                }
-                hasRecursiveTypeParameters
-              }
-              subst.lMap.get((tp.name, ScalaPsiUtil.getPsiElementId(tp.ptp))) match {
-                case Some(_addLower) =>
-                  val substedLowerType = unSubst.subst(lower)
-                  val addLower =
-                    if (tp.typeParams.nonEmpty && !_addLower.isInstanceOf[ScParameterizedType] &&
-                      !tp.typeParams.exists(_.name == "_"))
-                      ScParameterizedType(_addLower, tp.typeParams.map(ScTypeParameterType.toTypeParameterType))
-                    else _addLower
-                  if (hasRecursiveTypeParameters(substedLowerType)) lower = addLower
-                  else lower = Bounds.lub(substedLowerType, addLower)
-                case None =>
-                  lower = unSubst.subst(lower)
-              }
-              subst.rMap.get((tp.name, ScalaPsiUtil.getPsiElementId(tp.ptp))) match {
-                case Some(_addUpper) =>
-                  val substedUpperType = unSubst.subst(upper)
-                  val addUpper =
-                    if (tp.typeParams.nonEmpty && !_addUpper.isInstanceOf[ScParameterizedType] &&
-                      !tp.typeParams.exists(_.name == "_"))
-                      ScParameterizedType(_addUpper, tp.typeParams.map(ScTypeParameterType.toTypeParameterType))
-                    else _addUpper
-                  if (hasRecursiveTypeParameters(substedUpperType)) upper = addUpper
-                  else upper = Bounds.glb(substedUpperType, addUpper)
-                case None =>
-                  upper = unSubst.subst(upper)
-              }
-
-              if (safeCheck && !undefiningSubstitutor.subst(lower).conforms(undefiningSubstitutor.subst(upper), checkWeak = true))
-                throw new SafeCheckException
-              TypeParameter(tp.name, tp.typeParams /* doesn't important here */, () => lower, () => upper, tp.ptp)
-            }))
-          } else {
-            typeParams.foreach {case tp =>
-              val name = (tp.name, ScalaPsiUtil.getPsiElementId(tp.ptp))
-              if (un.names.contains(name)) {
-                def hasRecursiveTypeParameters(typez: ScType): Boolean = {
-                  var hasRecursiveTypeParameters = false
-                  typez.recursiveUpdate {
-                    case tpt: ScTypeParameterType =>
-                      typeParams.find(tp => (tp.name, ScalaPsiUtil.getPsiElementId(tp.ptp)) ==(tpt.name, tpt.getId)) match {
-                        case None => (true, tpt)
-                        case _ =>
-                          hasRecursiveTypeParameters = true
-                          (true, tpt)
-                      }
-                    case tp: ScType => (hasRecursiveTypeParameters, tp)
-                  }
-                  hasRecursiveTypeParameters
-                }
-                //todo: add only one of them according to variance
-                if (tp.lowerType() != Nothing) {
-                  val substedLowerType = unSubst.subst(tp.lowerType())
-                  if (!hasRecursiveTypeParameters(substedLowerType)) {
-                    un = un.addLower(name, substedLowerType, additional = true)
-                  }
-                }
-                if (tp.upperType() != Any) {
-                  val substedUpperType = unSubst.subst(tp.upperType())
-                  if (!hasRecursiveTypeParameters(substedUpperType)) {
-                    un = un.addUpper(name, substedUpperType, additional = true)
-                  }
-                }
-              }
-            }
-
-            def updateWithSubst(sub: ScSubstitutor): ScTypePolymorphicType = {
-              ScTypePolymorphicType(sub.subst(retType), typeParams.filter {
-                case tp =>
-                  val name = (tp.name, ScalaPsiUtil.getPsiElementId(tp.ptp))
-                  val removeMe: Boolean = un.names.contains(name)
-                  if (removeMe && safeCheck) {
-                    //let's check type parameter kinds
-                    def checkTypeParam(typeParam: ScTypeParam, tp: => ScType): Boolean = {
-                      val typeParams: Seq[ScTypeParam] = typeParam.typeParameters
-                      if (typeParams.isEmpty) return true
-                      tp match {
-                        case ScParameterizedType(_, typeArgs) =>
-                          if (typeArgs.length != typeParams.length) return false
-                          typeArgs.zip(typeParams).forall {
-                            case (tp: ScType, typeParam: ScTypeParam) => checkTypeParam(typeParam, tp)
-                          }
-                        case _ =>
-                          def checkNamed(named: PsiNamedElement, typeParams: Seq[ScTypeParam]): Boolean = {
-                            named match {
-                              case t: ScTypeParametersOwner =>
-                                if (typeParams.length != t.typeParameters.length) return false
-                                typeParams.zip(t.typeParameters).forall {
-                                  case (p1: ScTypeParam, p2: ScTypeParam) =>
-                                    if (p1.typeParameters.nonEmpty) checkNamed(p2, p1.typeParameters)
-                                    else true
-                                }
-                              case p: PsiTypeParameterListOwner =>
-                                if (typeParams.length != p.getTypeParameters.length) return false
-                                typeParams.forall(_.typeParameters.isEmpty)
-                              case _ => false
-                            }
-                          }
-                          ScType.extractDesignated(tp, withoutAliases = false) match {
-                            case Some((named, _)) => checkNamed(named, typeParams)
-                            case _ => tp match {
-                              case tpt: ScTypeParameterType => checkNamed(tpt.param, typeParams)
-                              case _ => false
-                            }
-                          }
-                      }
-                    }
-                    tp.ptp match {
-                      case typeParam: ScTypeParam =>
-                        if (!checkTypeParam(typeParam, sub.subst(new ScTypeParameterType(tp.ptp, ScSubstitutor.empty))))
-                          throw new SafeCheckException
-                      case _ =>
-                    }
-                  }
-                  !removeMe
-              }.map(tp => TypeParameter(tp.name, tp.typeParams /* doesn't important here */,
-                () => sub.subst(tp.lowerType()), () => sub.subst(tp.upperType()), tp.ptp)))
-            }
-
-            un.getSubstitutor match {
-              case Some(unSubstitutor) => updateWithSubst(unSubstitutor)
-              case _ if safeCheck => throw new SafeCheckException
-              case _ => updateWithSubst(unSubst)
-            }
-          }
-        case None => throw new SafeCheckException
-      }
-    } else ScTypePolymorphicType(retType, typeParams)
-    (tpe, c.problems, c.matchedArgs, c.matchedTypes)
   }
 
   def getElementsRange(start: PsiElement, end: PsiElement): Seq[PsiElement] = {
@@ -1263,7 +1050,7 @@ object ScalaPsiUtil {
 
   def typesCallSubstitutor(tp: Seq[(String, PsiElement)], typeArgs: Seq[ScType]): ScSubstitutor = {
     val map = new collection.mutable.HashMap[(String, PsiElement), ScType]
-    for (i <- 0 to math.min(tp.length, typeArgs.length) - 1) {
+    for (i <- 0 until math.min(tp.length, typeArgs.length)) {
       map += ((tp(i), typeArgs(i)))
     }
     new ScSubstitutor(Map(map.toSeq: _*), Map.empty, None)
@@ -1271,7 +1058,7 @@ object ScalaPsiUtil {
 
   def genericCallSubstitutor(tp: Seq[(String, PsiElement)], typeArgs: Seq[ScTypeElement]): ScSubstitutor = {
     val map = new collection.mutable.HashMap[(String, PsiElement), ScType]
-    for (i <- 0 to Math.min(tp.length, typeArgs.length) - 1) {
+    for (i <- 0 until Math.min(tp.length, typeArgs.length)) {
       map += ((tp(tp.length - 1 - i), typeArgs(typeArgs.length - 1 - i).getType(TypingContext.empty).getOrAny))
     }
     new ScSubstitutor(Map(map.toSeq: _*), Map.empty, None)
@@ -1520,8 +1307,8 @@ object ScalaPsiUtil {
             map { case (_, n) => n.info.asInstanceOf[PhysicalSignature] }
 
   def getMethodsForName(clazz: PsiClass, name: String): Seq[PhysicalSignature] = {
-    (for ((n: PhysicalSignature, _) <- TypeDefinitionMembers.getSignatures(clazz).forName(name)._1
-          if clazz.isInstanceOf[ScObject] || !n.method.hasModifierProperty("static")) yield n).toSeq
+    for ((n: PhysicalSignature, _) <- TypeDefinitionMembers.getSignatures(clazz).forName(name)._1
+         if clazz.isInstanceOf[ScObject] || !n.method.hasModifierProperty("static")) yield n
   }
 
   def getApplyMethods(clazz: PsiClass): Seq[PhysicalSignature] = {
@@ -1888,9 +1675,9 @@ object ScalaPsiUtil {
   def nthConstructorParam(cls: ScClass, n: Int): Option[ScParameter] = cls.constructor match {
     case Some(x: ScPrimaryConstructor) =>
       val clauses = x.parameterList.clauses
-      if (clauses.length == 0) None
+      if (clauses.isEmpty) None
       else {
-        val params = clauses(0).parameters
+        val params = clauses.head.parameters
         if (params.length > n)
           Some(params(n))
         else
@@ -1903,6 +1690,7 @@ object ScalaPsiUtil {
    * @return Some(parameter) if the expression is an argument expression that can be resolved to a corresponding
    *         parameter; None otherwise.
    */
+  @tailrec
   def parameterOf(exp: ScExpression): Option[Parameter] = {
     def forArgumentList(expr: ScExpression, args: ScArgumentExprList): Option[Parameter] = {
       args.getParent match {
@@ -2195,7 +1983,7 @@ object ScalaPsiUtil {
 
   def isViableForAssignmentFunction(fun: ScFunction): Boolean = {
     val clauses = fun.paramClauses.clauses
-    clauses.length == 0 || (clauses.length == 1 && clauses(0).isImplicit)
+    clauses.isEmpty || (clauses.length == 1 && clauses.head.isImplicit)
   }
 
   def padWithWhitespaces(element: PsiElement) {
@@ -2229,7 +2017,7 @@ object ScalaPsiUtil {
       case td: ScPatternDefinition => (td.bindings, td.expr)
       case _ => (Seq.empty[ScBindingPattern], None)
     }
-    if (bindings.size == 1 && expr.contains(instance)) Option(bindings(0))
+    if (bindings.size == 1 && expr.contains(instance)) Option(bindings.head)
     else {
       for (bind <- bindings) {
         if (bind.getType(TypingContext.empty).toOption == instance.getType(TypingContext.empty).toOption) return Option(bind)
@@ -2326,6 +2114,7 @@ object ScalaPsiUtil {
    * Should we check if it's a Single Abstract Method?
    * In 2.11 works with -Xexperimental
    * In 2.12 works by default
+ *
    * @return true if language level and flags are correct
    */
   def isSAMEnabled(e: PsiElement) = e.scalaLanguageLevel match {
@@ -2341,6 +2130,7 @@ object ScalaPsiUtil {
 
   /**
    * Determines if expected can be created with a Single Abstract Method and if so return the required ScType for it
+ *
    * @see SCL-6140
    * @see https://github.com/scala/scala/pull/3018/
    */
